side page components in separate folder
