+++ src/components/SidePageComponents/SidePageComponent.js
@@ -1,9 +1,9 @@
 import {
   EventHandlerService,
   PDFLEvents,
-} from "../services/EventHandlerService";
-import { KnowledgeGraphComponent } from "./KnowledgeGraphComponents/KnowledgeGraphComponent";
-import { SummaryKeyComponent } from "./SummaryKeyComponents/SummaryKeyComponent";
+} from "../../services/EventHandlerService";
+import { KnowledgeGraphComponent } from "../KnowledgeGraphComponents/KnowledgeGraphComponent";
+import { SummaryKeyComponent } from "../SummaryKeyComponents/SummaryKeyComponent";
 import { SidePageLoaderComponent } from "./SidePageLoaderComponent";
 
 /**
